refactor(cli): extract --env / --copy / --download / --volume parsers
The `Run`, `Exec`, and `pool run` subcommands all parsed the same
flag spec formats (`KEY=VAL`, `local:remote`, `remote:local`,
`name_or_path:guest`) inline. ~80 LOC of identical loops between
top-level `Run` and `pool run` alone, plus a smaller env-only
duplicate in `Exec`.

Extract four functions into a new `Sources/lumina-cli/SpecParsers.swift`:

  - `parseEnvSpecs(_:)`
  - `parseCopySpecs(_:)`            (returns the file/dir split)
  - `parseDownloadSpecs(_:)`
  - `parseVolumeSpecs(_:volumeStore:)`

Each writes the same friendly error to stderr and throws
`ExitCode.failure` on bad input — same shape as the inline
parsers they replace, so subcommand callers don't need their own
error formatting. Tests and live usage unchanged.

Net: -120 LOC of duplication across `Run`, `Exec`, and
`PoolCmd.Run`, +110 LOC of one shared file. ~10 LOC gross
shrink, but the win is in single-source-of-truth for the spec
formats — adding a new flag-bearing subcommand now wires four
helper calls instead of copying four loops.
