+++ Sources/lumina-cli/CLI.swift
@@ -105,84 +105,11 @@ struct Run: AsyncParsableCommand {
             throw ExitCode.failure
         }
 
-        var parsedEnv: [String: String] = [:]
-        for pair in env {
-            guard let eqIndex = pair.firstIndex(of: "=") else {
-                FileHandle.standardError.write(Data("lumina: invalid env '\(pair)'. Use KEY=VAL format\n".utf8))
-                throw ExitCode.failure
-            }
-            let key = String(pair[pair.startIndex..<eqIndex])
-            let value = String(pair[pair.index(after: eqIndex)...])
-            parsedEnv[key] = value
-        }
-
-        // Parse --copy: auto-detect file vs directory from local path
-        var parsedUploads: [FileUpload] = []
-        var parsedDirUploads: [DirectoryUpload] = []
-        for spec in copy {
-            guard let colonIndex = spec.firstIndex(of: ":") else {
-                FileHandle.standardError.write(Data("lumina: invalid --copy '\(spec)'. Use local:remote format\n".utf8))
-                throw ExitCode.failure
-            }
-            let localStr = String(spec[spec.startIndex..<colonIndex])
-            let remote = String(spec[spec.index(after: colonIndex)...])
-            let localURL = URL(fileURLWithPath: localStr)
-            var isDir: ObjCBool = false
-            guard FileManager.default.fileExists(atPath: localURL.path, isDirectory: &isDir) else {
-                FileHandle.standardError.write(Data("lumina: not found: \(localStr)\n".utf8))
-                throw ExitCode.failure
-            }
-            if isDir.boolValue {
-                parsedDirUploads.append(DirectoryUpload(localPath: localURL, remotePath: remote))
-            } else {
-                let mode = FileManager.default.isExecutableFile(atPath: localURL.path) ? "0755" : "0644"
-                parsedUploads.append(FileUpload(localPath: localURL, remotePath: remote, mode: mode))
-            }
-        }
-
-        // Parse --download: auto-detected as file vs directory at runtime on guest
-        var parsedDownloads: [FileDownload] = []
-        for spec in download {
-            guard let colonIndex = spec.firstIndex(of: ":") else {
-                FileHandle.standardError.write(Data("lumina: invalid --download '\(spec)'. Use remote:local format\n".utf8))
-                throw ExitCode.failure
-            }
-            let remote = String(spec[spec.startIndex..<colonIndex])
-            let localStr = String(spec[spec.index(after: colonIndex)...])
-            let localURL = URL(fileURLWithPath: localStr)
-            parsedDownloads.append(FileDownload(remotePath: remote, localPath: localURL))
-        }
-
-        // Parse --volume: host path (starts with / or .) = mount, otherwise = named volume
-        var parsedMounts: [MountPoint] = []
+        let parsedEnv = try parseEnvSpecs(env)
+        let (parsedUploads, parsedDirUploads) = try parseCopySpecs(copy)
+        let parsedDownloads = try parseDownloadSpecs(download)
         let volumeStore = VolumeStore()
-        for spec in volume {
-            guard let colonIndex = spec.firstIndex(of: ":") else {
-                FileHandle.standardError.write(Data("lumina: invalid --volume '\(spec)'. Use path_or_name:guest_path\n".utf8))
-                throw ExitCode.failure
-            }
-            let left = String(spec[spec.startIndex..<colonIndex])
-            let guestPath = String(spec[spec.index(after: colonIndex)...])
-
-            if left.hasPrefix("/") || left.hasPrefix(".") {
-                // Host directory mount
-                let hostURL = URL(fileURLWithPath: left)
-                var isDir: ObjCBool = false
-                guard FileManager.default.fileExists(atPath: hostURL.path, isDirectory: &isDir), isDir.boolValue else {
-                    FileHandle.standardError.write(Data("lumina: not a directory: \(left)\n".utf8))
-                    throw ExitCode.failure
-                }
-                parsedMounts.append(MountPoint(hostPath: hostURL, guestPath: guestPath))
-            } else {
-                // Named volume
-                guard let hostDir = volumeStore.resolve(name: left) else {
-                    FileHandle.standardError.write(Data("lumina: volume '\(left)' not found\n".utf8))
-                    throw ExitCode.failure
-                }
-                volumeStore.touch(name: left)
-                parsedMounts.append(MountPoint(hostPath: hostDir, guestPath: guestPath))
-            }
-        }
+        let parsedMounts = try parseVolumeSpecs(volume, volumeStore: volumeStore)
 
         var parsedDiskSize: UInt64? = nil
         if let ds = resolvedDiskSize {
@@ -1278,14 +1205,7 @@ struct Exec: AsyncParsableCommand {
         }
         let timeoutSecs = Int(parsedTimeout.components.seconds)
 
-        var parsedEnv: [String: String] = [:]
-        for pair in env {
-            guard let eqIndex = pair.firstIndex(of: "=") else {
-                FileHandle.standardError.write(Data("lumina: invalid env '\(pair)'. Use KEY=VAL\n".utf8))
-                throw ExitCode.failure
-            }
-            parsedEnv[String(pair[..<eqIndex])] = String(pair[pair.index(after: eqIndex)...])
-        }
+        let parsedEnv = try parseEnvSpecs(env)
 
         let client = SessionClient()
         do {
@@ -1950,78 +1870,11 @@ struct PoolRun: AsyncParsableCommand {
             FileHandle.standardError.write(Data("lumina: invalid memory: \(memory)\n".utf8))
             throw ExitCode.failure
         }
-        var parsedEnv: [String: String] = [:]
-        for pair in env {
-            guard let eqIndex = pair.firstIndex(of: "=") else {
-                FileHandle.standardError.write(Data("lumina: invalid env '\(pair)'. Use KEY=VAL format\n".utf8))
-                throw ExitCode.failure
-            }
-            parsedEnv[String(pair[pair.startIndex..<eqIndex])] = String(pair[pair.index(after: eqIndex)...])
-        }
-
-        // Parse --copy: auto-detect file vs directory from local path
-        var parsedUploads: [FileUpload] = []
-        var parsedDirUploads: [DirectoryUpload] = []
-        for spec in copy {
-            guard let colonIndex = spec.firstIndex(of: ":") else {
-                FileHandle.standardError.write(Data("lumina: invalid --copy '\(spec)'. Use local:remote format\n".utf8))
-                throw ExitCode.failure
-            }
-            let localStr = String(spec[spec.startIndex..<colonIndex])
-            let remote = String(spec[spec.index(after: colonIndex)...])
-            let localURL = URL(fileURLWithPath: localStr)
-            var isDir: ObjCBool = false
-            guard FileManager.default.fileExists(atPath: localURL.path, isDirectory: &isDir) else {
-                FileHandle.standardError.write(Data("lumina: not found: \(localStr)\n".utf8))
-                throw ExitCode.failure
-            }
-            if isDir.boolValue {
-                parsedDirUploads.append(DirectoryUpload(localPath: localURL, remotePath: remote))
-            } else {
-                let mode = FileManager.default.isExecutableFile(atPath: localURL.path) ? "0755" : "0644"
-                parsedUploads.append(FileUpload(localPath: localURL, remotePath: remote, mode: mode))
-            }
-        }
-
-        // Parse --download: auto-detected as file vs directory at runtime on guest
-        var parsedDownloads: [FileDownload] = []
-        for spec in download {
-            guard let colonIndex = spec.firstIndex(of: ":") else {
-                FileHandle.standardError.write(Data("lumina: invalid --download '\(spec)'. Use remote:local format\n".utf8))
-                throw ExitCode.failure
-            }
-            let remote = String(spec[spec.startIndex..<colonIndex])
-            let localStr = String(spec[spec.index(after: colonIndex)...])
-            parsedDownloads.append(FileDownload(remotePath: remote, localPath: URL(fileURLWithPath: localStr)))
-        }
-
-        // Parse --volume: applied at pool boot time (VM-level config)
-        var parsedMounts: [MountPoint] = []
+        let parsedEnv = try parseEnvSpecs(env)
+        let (parsedUploads, parsedDirUploads) = try parseCopySpecs(copy)
+        let parsedDownloads = try parseDownloadSpecs(download)
         let volumeStore = VolumeStore()
-        for spec in volume {
-            guard let colonIndex = spec.firstIndex(of: ":") else {
-                FileHandle.standardError.write(Data("lumina: invalid --volume '\(spec)'. Use path_or_name:guest_path\n".utf8))
-                throw ExitCode.failure
-            }
-            let left = String(spec[spec.startIndex..<colonIndex])
-            let guestPath = String(spec[spec.index(after: colonIndex)...])
-            if left.hasPrefix("/") || left.hasPrefix(".") {
-                let hostURL = URL(fileURLWithPath: left)
-                var isDir: ObjCBool = false
-                guard FileManager.default.fileExists(atPath: hostURL.path, isDirectory: &isDir), isDir.boolValue else {
-                    FileHandle.standardError.write(Data("lumina: not a directory: \(left)\n".utf8))
-                    throw ExitCode.failure
-                }
-                parsedMounts.append(MountPoint(hostPath: hostURL, guestPath: guestPath))
-            } else {
-                guard let hostDir = volumeStore.resolve(name: left) else {
-                    FileHandle.standardError.write(Data("lumina: volume '\(left)' not found\n".utf8))
-                    throw ExitCode.failure
-                }
-                volumeStore.touch(name: left)
-                parsedMounts.append(MountPoint(hostPath: hostDir, guestPath: guestPath))
-            }
-        }
+        let parsedMounts = try parseVolumeSpecs(volume, volumeStore: volumeStore)
 
         let opts = VMOptions(memory: parsedMemory, cpuCount: cpus, image: image, mounts: parsedMounts)
